refactor: CategoryField 분리
